# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/albums": {
            "get": {
                "description": "Returns all photo albums as a flat list. Use ?tree=true to get a nested tree.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "albums"
                ],
                "summary": "List all photo albums",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Return as nested tree (default false)",
                        "name": "tree",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/v0_albums.AlbumJSON"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new photo album, optionally nested under a parent.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "albums"
                ],
                "summary": "Create a photo album",
                "parameters": [
                    {
                        "description": "Album name and optional parent ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v0_albums.createAlbumRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/v0_albums.AlbumJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            }
        },
...
        "/albums/{id}/children": {
            "post": {
                "description": "Shorthand for creating an album with a specific parent ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "albums"
                ],
                "summary": "Create a child album under a parent",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Parent album ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Album name",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v0_albums.createAlbumRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/v0_albums.AlbumJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            }
        },
...
        "/cirrus/download": {
            "get": {
                "description": "Downloads a single file or zips a folder and streams it back to the client",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "cirrus"
                ],
                "summary": "Download a file or folder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File path to download",
                        "name": "filePath",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Device serial number to filter by",
                        "name": "serial",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Output format conversion (e.g. 'jpeg' to convert HEIC to JPEG)",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "serverutil.ContentType": {
            "type": "string",
            "enum": [
                "text/html",
                "application/json"
            ],
            "x-enum-varnames": [
                "ContentTypeHTML",
                "ContentTypeJSON"
            ]
        },
        "serverutil.Response": {
            "type": "object",
            "properties": {
                "contentType": {
                    "$ref": "#/definitions/serverutil.ContentType"
                },
                "data": {},
                "error": {},
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
        "v0_albums.AlbumJSON": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v0_albums.AlbumJSON"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "itemCount": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "integer"
                },
                "smartType": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "v0_albums.createAlbumRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "integer"
                }
            }
        },
...
    }
}